Validate error codes.
Earlier today I tried to print but the printer was low on battery.  I got a value error.

Bad user experience.

Now we look up which error code the printer returned, and return a PrinterError exception containing it.  This then gets turned into a HomeAssistantError (as per the HA docs) so the message can be shown to the user.