Validate error codes. (#20)
* Configurable and testable performance options.

* Image preview generation.

README has the nitty gritty.

I deleted the code that writes images to disk.  It was... less than ideal.

* Add an image entity to the label printer.

* new_multiline: Auto-fit both width and height.

* Remove stray prints.

* Update preview image entity also when printing.

* Add options flow.

* Validate error codes.

Earlier today I tried to print but the printer was low on battery.  I got a value error.

Bad user experience.

Now we look up which error code the printer returned, and return a PrinterError exception containing it.  This then gets turned into a HomeAssistantError (as per the HA docs) so the message can be shown to the user.

---------

Co-authored-by: eigger <eigger87@gmail.com>